Remove Microsoft from project namespace name.
