--- Microsoft.EntityFrameworkCore.UnitOfWork.Host/Models/CustomBlogRepository.cs
@@ -1,10 +1,0 @@
-﻿namespace Microsoft.EntityFrameworkCore.UnitOfWork.Host.Models
-{
-    public class CustomBlogRepository : Repository<Blog>, IRepository<Blog>
-    {
-        public CustomBlogRepository(BloggingContext dbContext) : base(dbContext)
-        {
-
-        }
-    }
-}
