--- Microsoft.EntityFrameworkCore.UnitOfWork.Tests/Entities/City.cs
@@ -1,17 +1,0 @@
-﻿using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations.Schema;
-
-namespace Microsoft.EntityFrameworkCore.UnitOfWork.Tests.Entities
-{
-    public class City
-    {
-        public int Id { get; set; }
-        public string Name { get; set; }
-        
-        public int CountryId { get; set; }
-
-        public Country Country { get; set; }
-
-        public List<Town> Towns { get; set; }
-    }
-}
