--- Microsoft.EntityFrameworkCore.UnitOfWork.Tests/Entities/Customer.cs
@@ -1,9 +1,0 @@
-﻿namespace Microsoft.EntityFrameworkCore.UnitOfWork.Tests.Entities
-{
-    public class Customer
-    {
-        public int Id { get; set; }
-        public string Name { get; set; }
-        public int Age { get; set; }
-    }
-}
