--- Microsoft.EntityFrameworkCore.UnitOfWork/IUnitOfWorkOfT.cs
@@ -1,25 +1,0 @@
-﻿// Copyright (c) Arch team. All rights reserved.
-
-using System.Threading.Tasks;
-
-namespace Microsoft.EntityFrameworkCore.UnitOfWork
-{
-    /// <summary>
-    /// Defines the interface(s) for generic unit of work.
-    /// </summary>
-    public interface IUnitOfWork<TContext> : IUnitOfWork where TContext : DbContext {
-        /// <summary>
-        /// Gets the db context.
-        /// </summary>
-        /// <returns>The instance of type <typeparamref name="TContext"/>.</returns>
-        TContext DbContext { get; }
-
-        /// <summary>
-        /// Saves all changes made in this context to the database with distributed transaction.
-        /// </summary>
-        /// <param name="ensureAutoHistory"><c>True</c> if save changes ensure auto record the change history.</param>
-        /// <param name="unitOfWorks">An optional <see cref="IUnitOfWork"/> array.</param>
-        /// <returns>A <see cref="Task{TResult}"/> that represents the asynchronous save operation. The task result contains the number of state entities written to database.</returns>
-        Task<int> SaveChangesAsync(bool ensureAutoHistory = false, params IUnitOfWork[] unitOfWorks);
-    }
-}
