+++ UnitOfWork.Host/Models/CustomBlogRepository.cs
@@ -1,0 +1,10 @@
+﻿namespace Arch.EntityFrameworkCore.UnitOfWork.Host.Models
+{
+    public class CustomBlogRepository : Repository<Blog>, IRepository<Blog>
+    {
+        public CustomBlogRepository(BloggingContext dbContext) : base(dbContext)
+        {
+
+        }
+    }
+}
