+++ UnitOfWork.Tests/Entities/Country.cs
@@ -1,0 +1,12 @@
+﻿using System.Collections.Generic;
+
+namespace Arch.EntityFrameworkCore.UnitOfWork.Tests.Entities
+{
+    public class Country
+    {
+        public int Id { get; set; }
+        public string Name { get; set; }
+
+        public List<City> Cities { get; set; }
+    }
+}
