+++ UnitOfWork.Tests/Entities/Customer.cs
@@ -1,0 +1,9 @@
+﻿namespace Arch.EntityFrameworkCore.UnitOfWork.Tests.Entities
+{
+    public class Customer
+    {
+        public int Id { get; set; }
+        public string Name { get; set; }
+        public int Age { get; set; }
+    }
+}
