+++ UnitOfWork/UnitOfWork.csproj
@@ -1,0 +1,22 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+  <PropertyGroup>
+    <Description>A plugin for Microsoft.EntityFrameworkCore to support repository, unit of work patterns, and multiple database with distributed transaction supported.</Description>
+    <VersionPrefix>2.1.0</VersionPrefix>
+    <Authors>rigofunc;rigofunc@outlook.com;</Authors>
+    <TargetFramework>netstandard2.0</TargetFramework>
+    <NoWarn>$(NoWarn);CS1591</NoWarn>
+    <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
+    <GenerateDocumentationFile>true</GenerateDocumentationFile>
+    <AssemblyName>Arch.EntityFrameworkCore.UnitOfWork</AssemblyName>
+    <PackageId>Arch.EntityFrameworkCore.UnitOfWork</PackageId>
+    <PackageTags>Entity Framework Core;entity-framework-core;EF;Data;O/RM;unitofwork;Unit Of Work;unit-of-work</PackageTags>
+    <PackageProjectUrl>https://github.com/arch/UnitOfWork</PackageProjectUrl>
+    <PackageLicenseUrl>https://github.com/arch/UnitOfWork/blob/master/LICENSE</PackageLicenseUrl>
+    <RepositoryType>git</RepositoryType>
+    <RepositoryUrl>https://github.com/arch/UnitOfWork.git</RepositoryUrl>
+  </PropertyGroup>
+  <ItemGroup>
+    <PackageReference Include="Microsoft.EntityFrameworkCore.AutoHistory" Version="2.1.0" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore.Relational" Version="2.1.2" />
+  </ItemGroup>
+</Project>
